KuMo

Documentation for KuMo.

KuMo.AbstractJob

KuMo.AbstractResource

KuMo.Data

KuMo.Job

KuMo.Topology

KuMo.graph

KuMo.job

KuMo.rand_job

KuMo.vtx

KuMo.AbstractJob — Type

AbstractJob

An abstract supertype for jobs.

KuMo.AbstractResource — Type

AbstractResource

An abstract supertype for resources in a cloud morphing architecture. Any type MyResource <: AbstractResource needs to either:

have a field capacity::T where T <: Number,

implement a capacity(r::MyResource) method.

Optionally, one can implement a specific pseudo_cost(r::MyResource, charge) method.

KuMo.Data — Type

Data

Structure to store the information related to some Data. Currently, only the location of such data is stored.

KuMo.Job — Type

Job <: AbstractJob

The most generic job type.

KuMo.Topology — Type

Topology{N<:AbstractNode,L<:AbstractLink}

A structure to store the topology of a network. Beside the graph structure itself, it also stores the kinds of all nodes and links.

KuMo.graph — Method

graph(topo::Topology, algorithm::AbstractAlgorithm)

Creates an appropriate digraph using Graph.jl based on a topology and the requirement of an algorithm.

KuMo.job — Method

job(backend::Int, containers::Int, data_location::Int, duration::Float64, frontend::Int)

Method to create new jobs.

KuMo.rand_job — Method

rand_job(jd::Dict)

Create a random job given a job_distribution dictionary.

KuMo.vtx — Method

vtx(algorithm::AbstractAlgorithm)

Return the number of additional vertices required by the algorithm used to allocate resources in the network.