KuMo

Documentation for KuMo.

KuMo.SCENARII

KuMo.AbstractAlgorithm

KuMo.AbstractJob

KuMo.AbstractRequests

KuMo.AbstractResource

KuMo.AdditiveNode

KuMo.ConvexLink

KuMo.Data

KuMo.DirectedTopology

KuMo.EqualLoadBalancingNode

KuMo.FlatLink

KuMo.FlatNode

KuMo.FreeLink

KuMo.IdleStateNode

KuMo.Job

KuMo.Link

KuMo.Load

KuMo.MultiplicativeNode

KuMo.Node

KuMo.PeriodicRequests

KuMo.PremiumNode

KuMo.Request

KuMo.Requests

KuMo.Scenario

KuMo.ShortestPath

KuMo.SnapShot

KuMo.State

KuMo.Topology

KuMo.Unload

KuMo.User

KuMo.add_load!

KuMo.capacity

KuMo.clean

KuMo.cond_minmax

KuMo.data

KuMo.execute_valid_load

KuMo.graph

KuMo.init_simulate

KuMo.init_simulate

KuMo.init_simulate

KuMo.init_user

KuMo.init_user

KuMo.inner_queue

KuMo.insert_sorted!

KuMo.job

KuMo.job_distributions

KuMo.location

KuMo.make_df

KuMo.make_df

KuMo.make_links

KuMo.make_nodes

KuMo.make_users

KuMo.marks

KuMo.param

KuMo.post_simulate

KuMo.pseudo_cost

KuMo.pseudo_cost

KuMo.push_snap!

KuMo.rand_job

KuMo.rem_load!

KuMo.requests

KuMo.requests

KuMo.requests

KuMo.retrieve_path

KuMo.scenario

KuMo.simulate

KuMo.simulate_loop

KuMo.simulate_loop

KuMo.smooth

KuMo.spike

KuMo.steady

KuMo.user

KuMo.vtx

KuMo.SCENARII — Constant

SCENARII

Collection of scenarii.

KuMo.AbstractAlgorithm — Type

AbstractAlgorithm

An abstract supertype for algorithms.

KuMo.AbstractJob — Type

AbstractJob

An abstract supertype for jobs.

KuMo.AbstractRequests — Type

AbstractRequests

An abstract supertype for job requests.

KuMo.AbstractResource — Type

AbstractResource

An abstract supertype for resources in a cloud morphing architecture. Any type MyResource <: AbstractResource needs to either:

have a field capacity::T where T <: Number,

implement a capacity(r::MyResource) method.

Optionally, one can implement a specific pseudo_cost(r::MyResource, charge) method.

KuMo.AdditiveNode — Type

AdditiveNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure where the default pseudo-cost is translated by the value in the param field.

KuMo.ConvexLink — Type

ConvexLink <: KuMo.AbstractLink

Link structure with a convex pseudo-cost function.

KuMo.Data — Type

Data

Structure to store the information related to some Data. Currently, only the location of such data is stored.

KuMo.DirectedTopology — Type

DirectedTopology{N<:AbstractNode,L<:AbstractLink}

A structure to store the topology of a network. Beside the graph structure itself, it also stores the kinds of all nodes and links.

KuMo.EqualLoadBalancingNode — Type

EqualLoadBalancingNode{T <: Number} <: AbstractNode

Node structure with an equal load balancing (monotonic) pseudo-cost function.

KuMo.FlatLink — Type

FlatLink <: KuMo.AbstractLink

Link structure with a constant pseudo-cost function.

KuMo.FlatNode — Type

FlatNode{T <: Number} <: AbstractNode

Node structure with a constant pseudo-cost function.

KuMo.FreeLink — Type

FreeLink <: AbstractLink

The pseudo-cost of such links is always zero.

KuMo.IdleStateNode — Type

IdleStateNode{T1 <: Number, T2 <: Number} <: AbstractNode

Node structure that stays iddle until a bigger system load than the default node. The param field is used to set the activation treshold.

KuMo.Job — Type

Job <: AbstractJob

The most generic job type.

Arguments:

backend::Int: size of the backend data to be sent from data location to the server

containers::Int: number of containers required to execute the job

data_location::Int: location of the data (node id)

duration::Float64: job duration

frontend::Int: size of the frontend data to be sent from the user location to the server

KuMo.Link — Type

Link{T <: Number} <: AbstractLink

Default link structure with an equal load balancing (monotonic) pseudo-cost function.

KuMo.Load — Type

Load

A structure describing an increase of the total load.

Arguments:

occ::Float64: time when the load occurs

node::Int: node at which the request is executed

job::Job: the job request

KuMo.MultiplicativeNode — Type

MultiplicativeNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure where the default pseudo-cost is multiplied by the value in the param field.

KuMo.Node — Type

Node{T <: Number} <: AbstractNode

Default node structure, defined by its maximal capacity and the default convex pseudo-cost function.

KuMo.PeriodicRequests — Type

PeriodicRequests{J < :AbstractJob} <: AbstractRequests

A structure to handle job that

Arguments:

job::J: the job being requested periodically

period::Float64

start::Float64

stop::Float64

`PeriodicRequests(j, p; start = -Inf, stop = Inf): default constructor

KuMo.PremiumNode — Type

PremiumNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure for premium resources. The param field set the premium treshold.

KuMo.Request — Type

Request{J <: AbstractJob}

Single unrepeated request.

Arguments:

job::J: the job being requested periodically

start::Float64

KuMo.Requests — Type

Requests{J <: AbstractJob} <: AbstractRequests

A collection of aperiodic requests.

KuMo.Scenario — Type

Scenario{N <: AbstractNode, L <: AbstractLink, U <: User}

Structure to store the information of a scenario.

Arguments:

data::Dictionary{Int, Data}: data collection (currently not in use)

duration::Real: optional duration of the scenario

topology::Topology{N, L}: network's topology

users::Dictionary{Int, U}: collection of users

KuMo.ShortestPath — Type

ShortestPath <: AbstractAlgorithm

A ShortestPath algorithm.

KuMo.SnapShot — Type

SnapShot

A structure to take snapshot from the state of network and its resources at a specific instant.

Arguments:

state::State: state at instant

total::Float64: total load at instant

selected::Int: selected node at instant; value is zero if load is removed

duration::Float64: duration of all the actions taken during corresponding to the state of this snap

solving_time::Float64: time taken specifically by the solving algorithm (<: AbstractAlgorithm)

instant::Float64

KuMo.State — Type

State

A structure to store the state of the different resources, e.g. nodes and links, during a simulation.

Arguments:

links::SparseMatrixCSC{Float64, Int64}: sparse matrice with the links loads

nodes::SparseVector{Float64, Int64}: sparse vector with the nodes loads

State(n): inner constructor given the number of nodes n

State(links, nodes): inner constructor given the links and nodes of an existing state

KuMo.Topology — Type

Topology{N<:AbstractNode,L<:AbstractLink}

A structure to store the topology of a network. Beside the graph structure itself, it also stores the kinds of all nodes and links.

KuMo.Unload — Type

Unload

A structure describing a decrease of the total load.

Arguments:

occ::Float64: time when the unload occurs

node::Int: node at which the request was executed

vload::Int: the number of freed containers

lloads::SparseMatrixCSC{Float64, Int64}: the freed loads on each link

KuMo.User — Type

User{R<:AbstractRequests}

A structure to store a user information. A user is defined through a sequence of requests and a location (node id).

KuMo.add_load! — Method

add_load!(state, links, containers, v, n)

Adds load to a given state.

Arguments:

state

links: the load increase to be added on links

containers: the containers load to be added to v

v: node selected to execute a task

n: amount of available nodes

KuMo.capacity — Method

capacity(r::R) where {R<:AbstractResource}

Return the capacity of a resource r.

KuMo.clean — Method

clean(snaps)

Clean the snapshots by merging snaps occuring at the same time.

KuMo.cond_minmax — Method

cond_minmax(x, y, b)

Return the min of x and y if b is false, and the max of x and y otherwise.

KuMo.data — Function

data(location::Int = 0)

Construct data at location

KuMo.execute_valid_load — Function

execute_valid_load(s, task, g, capacities, state, algo, demands, ii = 0)

Compute the best load allocation and return if it is a valid one.

Arguments:

s: scenario being simulated

task: task being requested

g: graph of the network topology

capacities: capacities of the network

state: current state of resources

algo: algo used for computing the best allocation cost

demands: if algo is KuMoFlowExt.MinCostFlow, demands are required

ii: a counter to measure the approximative progress of the simulation

KuMo.graph — Method

graph(topo::Topology, algorithm::AbstractAlgorithm)

Creates an appropriate digraph using Graph.jl based on a topology and the requirement of an algorithm.

KuMo.init_simulate — Method

init_simulate(s, algo, tasks, start)

Initialize structures before the simualtion of scenario s.

Arguments:

s: the scenario being simulated

algo: algorithm allocating resources at best known lower costs resources

tasks: sorted container of tasks to be simulated

start: instant when the simulation started

KuMo.init_simulate — Method

init_simulate(::Val{0})

Initialize a synchronous simulation.

KuMo.init_simulate — Method

init_simulate(::Val)

Initialize an asynchronous simulation.

KuMo.init_user — Method

init_user(s::Scenario, u::User, tasks, ::PeriodicRequests)

Initialize user u periodic requests.

Arguments:

s: scenario that is about to be simulated

u: a user id

tasks: container of sorted tasks

KuMo.init_user — Method

init_user(::Scenario, u::User, tasks, ::Requests)

Initialize user u non-periodic requests.

Arguments:

u: a user id

tasks: container of sorted tasks

KuMo.inner_queue — Function

inner_queue(g, u, j, nodes, capacities, state, ::ShortestPath, ii = 0; lck = ReentrantLock(), demands = nothing, links)

DOCSTRING

Arguments:

g: a graph representing the topology of the network

u: user location

j: requested job

nodes: nodes capacities

capacities: links capacities

state: current state of the network

algo: ShortestPath <: AbstractAlgorithm

ii: a counter to mesure the progress in the simulation

lck: a lck for asynchronous simulation

demands: not needed for ShortestPath algorithm

links: description of the links topology

KuMo.insert_sorted! — Function

insert_sorted!(w, val, it = iterate(w))

Insert element in a sorted collection.

Arguments:

w: sorted collection

val: value to be inserted

it: optional iterator

KuMo.job — Method

job(backend::Int, containers::Int, data_location::Int, duration::Float64, frontend::Int)

Method to create new jobs.

KuMo.job_distributions — Method

job_distributions(; backend, container, data_locations, duration, frontend)

Construct a dictionary with random distributions to generate new jobs. Beside data_locations, the other arguments should be a 2-tuple defining normal distributions as in the Distributions.jl package.

Arguments:

backend

container

data_locations: a collection/range of possible data location

duration

frontend

KuMo.location — Method

location(d::Data)

Returns the location of the data d.

KuMo.make_df — Method

make_df(s::Scenario; verbose = true)

Make a DataFrame to describe the scenario s.

KuMo.make_df — Method

make_df(snapshots::Vector{SnapShot}, topo; verbose = true)

Make a DataFrame from the raw snapshots.

Arguments:

snapshots: A collection of snapshots

topo: topology of the network

verbose: if set to true, it will print a description of the snapshots in the terminal

KuMo.make_links — Method

make_links(links)

Creates links.

KuMo.make_nodes — Method

make_nodes(nodes)

Create nodes.

KuMo.make_users — Method

make_users(args...)

Create users.

KuMo.marks — Method

marks(df::DataFrame)

Returns a 4-tuple (a, b, c, d) that marks the start and end of the nodes and links columns in the dataframe.

(a, b) mark the start and end indices of the nodes columns

(c, d) mark the start and end indices of the links columns

KuMo.param — Method

param(r::R) where {R<:AbstractResource}

Default accessor for an optional parameter for R. If no param field exists, returns nothing.

KuMo.post_simulate — Method

post_simulate(s, snapshots, verbose, output)

Post-simulation process that covers cleaning the snapshots and producing an output.

Arguments:

s: simulated scenario

snapshots: resulting snapshots (before cleaning)

verbose: if set to true, prints information about the output and the snapshots

output: output path

KuMo.pseudo_cost — Function

pseudo_cost(cap, charge, resource, param...)
pseudo_cost(r::<:AbstractResource, charge)

Methods to compute the pseudo-cost of various resources.

KuMo.pseudo_cost — Method

pseudo_cost(r::R, charge) where {R<:AbstractResource}

Compute the pseudo-cost of r given its charge.

KuMo.push_snap! — Method

push_snap!(snapshots, state, total, selected, duration, solving_time, instant, n)

Add a snapshot to an existing collection of snapshots.

Arguments:

snapshots: collection of snapshots

state: current state

total: load

selected: node where a request is executed

duration: duration of the whole resource allocation for the request

solving_time: time taken specifically by the solving algorithm (<: AbstractAlgorithm)

instant: instant when the request is received

n: number of available nodes

KuMo.rand_job — Method

rand_job(jd::Dict)

Create a random job given a job_distribution dictionary.

KuMo.rem_load! — Method

rem_load!(state, links, containers, v, n)

Removes load from a given state.

Arguments:

state

links: the load increase to be removed from links

containers: the containers load to be removed from v

v: node where a task is endind

n: amount of available nodes

KuMo.requests — Method

requests(requests_lst...)

Construct a collection of requests.

KuMo.requests — Method

requests(j::Job, n::Int, d::UnivariateDistribution, lower::Real, upper::Real)

Generate a sequence of n requests with the same job following the distribution d. Limits, lower and upper, can be specified to truncate d.

KuMo.requests — Method

requests(pr::PeriodicRequests)

Generate a sequence of aperiodic requests from a periodic request pr.

KuMo.retrieve_path — Method

retrieve_path(u, v, paths)

Retrieves the path from u to v.

Arguments:

u: source vertex

v: target vertex

paths: list of shortest paths within a network

KuMo.scenario — Method

scenario(; duration, links = nothing, nodes, users, job_distribution = nothing, request_rate = nothing)

Build a scenario.

Arguments:

duration: duration of the interval where requests can be started

links: collection of links resources

nodes: collection of nodes resources

users: collections of users information

job_distribution: (optional) distributions used to generate jobs

request_rate: (optional) average request rate

KuMo.simulate — Function

simulate(s::Scenario, algo; speed = 0, output = "", verbose = true)

Simulate a scenario.

Arguments:

s: simulation targetted scenario

algo: algorithm used to estimate the best allocation regarding to the pseudo-cost

speed: simulation speed. If set to 0, the requests are handled sequentially without computing time limits. Otherwise the requests are made as independant asynchronous processes

output: path to save the output, if empty (default), nothing is saved

verbose: if set to true, prints information about the simulation

KuMo.simulate_loop — Method

simulate_loop(s, algo, _, start, containers, args_loop, ::Val{0})

Inner loop of the simulation of scenario s.

Arguments:

s: scenario being simulated

algo: algo solving the resource allocation dynamically at each step

_: simulation speed (unrequired)

start: starting time of the simulation

containers: containers generated to allocate tasks dynamically during the run

args_loop: arguments required by this loop

KuMo.simulate_loop — Method

simulate_loop(s, algo, speed, start, containers, args_loop, ::Val)

Inner loop of the simulation of scenario s.

Arguments:

s: scenario being simulated

algo: algo solving the resource allocation dynamically at each step

speed: asynchronous simulation speed

start: starting time of the simulation

containers: containers generated to allocate tasks dynamically during the run

args_loop: arguments required by this loop

KuMo.smooth — Method

smooth(j, δ, π1, π2)

Generate a collection of requests for job j in [π1, π2] that grows smoothly in intensity. Requests are emitted every δ interval.

KuMo.spike — Method

spike(j, t, intensity)

Generate a spike of requests for job j at instant t. The number of requests is defined by intensity.

KuMo.steady — Method

steady(j, δ, π1, π2, intensity)

Generate a collection of requests for job j in [π1, π2] with a constant intensity. Requests are emitted every δ interval.

KuMo.user — Function

user(job_distributions::Dict, period, loc; start=-Inf, stop=Inf)
user(job, period, loc; start=-Inf, stop=Inf)
user(jr::Vector{R}, loc::Int) where {R<:Request}
user(jr, loc::Int)
user(jr, loc)

A serie of methods to generate users.

KuMo.vtx — Method

vtx(algorithm::AbstractAlgorithm)

Return the number of additional vertices required by the algorithm used to allocate resources in the network.