KuMo

Documentation for KuMo.

KuMo.AbstractAlgorithm

KuMo.AbstractJob

KuMo.AbstractRequest

KuMo.AbstractResource

KuMo.AdditiveNode

KuMo.ConvexLink

KuMo.Data

KuMo.DirectedTopology

KuMo.EqualLoadBalancingNode

KuMo.FlatLink

KuMo.FlatNode

KuMo.FreeLink

KuMo.IdleStateNode

KuMo.Job

KuMo.Link

KuMo.MultiplicativeNode

KuMo.Node

KuMo.PremiumNode

KuMo.ShortestPath

KuMo.SnapShot

KuMo.State

KuMo.Topology

KuMo.User

KuMo.add_load!

KuMo.capacity

KuMo.clean

KuMo.cond_minmax

KuMo.data

KuMo.execution_results

KuMo.execution_results

KuMo.graph

KuMo.init_execution

KuMo.init_execution

KuMo.inner_queue

KuMo.insert_sorted!

KuMo.job

KuMo.job_distributions

KuMo.make_df

KuMo.marks

KuMo.param

KuMo.pseudo_cost

KuMo.pseudo_cost

KuMo.push_snap!

KuMo.rand_job

KuMo.rem_load!

KuMo.retrieve_path

KuMo.user

KuMo.valid_load

KuMo.vtx

KuMo.AbstractAlgorithm — Type

AbstractAlgorithm

An abstract supertype for algorithms.

KuMo.AbstractJob — Type

AbstractJob

An abstract supertype for jobs.

KuMo.AbstractRequest — Type

AbstractRequests

An abstract supertype for job requests.

KuMo.AbstractResource — Type

AbstractResource

An abstract supertype for resources in a cloud morphing architecture. Any type MyResource <: AbstractResource needs to either:

have a field capacity::T where T <: Number,

implement a capacity(r::MyResource) method.

Optionally, one can implement a specific pseudo_cost(r::MyResource, charge) method.

KuMo.AdditiveNode — Type

AdditiveNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure where the default pseudo-cost is translated by the value in the param field.

KuMo.ConvexLink — Type

ConvexLink <: KuMo.AbstractLink

Link structure with a convex pseudo-cost function.

KuMo.Data — Type

Data

Structure to store the information related to some Data. Currently, only the location of such data is stored.

KuMo.DirectedTopology — Type

DirectedTopology{N<:AbstractNode,L<:AbstractLink}

A structure to store the topology of a network. Beside the graph structure itself, it also stores the kinds of all nodes and links.

KuMo.EqualLoadBalancingNode — Type

EqualLoadBalancingNode{T <: Number} <: AbstractNode

Node structure with an equal load balancing (monotonic) pseudo-cost function.

KuMo.FlatLink — Type

FlatLink <: KuMo.AbstractLink

Link structure with a constant pseudo-cost function.

KuMo.FlatNode — Type

FlatNode{T <: Number} <: AbstractNode

Node structure with a constant pseudo-cost function.

KuMo.FreeLink — Type

FreeLink <: AbstractLink

The pseudo-cost of such links is always zero.

KuMo.IdleStateNode — Type

IdleStateNode{T1 <: Number, T2 <: Number} <: AbstractNode

Node structure that stays iddle until a bigger system load than the default node. The param field is used to set the activation threshold.

KuMo.Job — Type

Job <: AbstractJob

The most generic job type.

Arguments:

backend::Int: size of the backend data to be sent from data location to the server

containers::Int: number of containers required to execute the job

duration::Float64: job duration

frontend::Int: size of the frontend data to be sent from the user location to the server

KuMo.Link — Type

Link{T <: Number} <: AbstractLink

Default link structure with an equal load balancing (monotonic) pseudo-cost function.

KuMo.MultiplicativeNode — Type

MultiplicativeNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure where the default pseudo-cost is multiplied by the value in the param field.

KuMo.Node — Type

Node{T <: Number} <: AbstractNode

Default node structure, defined by its maximal capacity and the default convex pseudo-cost function.

KuMo.PremiumNode — Type

PremiumNode{T1 <: Number, T2 <: Number} <: AbstractNode

A node structure for premium resources. The param field set the premium threshold.

KuMo.ShortestPath — Type

ShortestPath <: AbstractAlgorithm

A ShortestPath algorithm.

KuMo.SnapShot — Type

SnapShot

A structure to take snapshot from the state of network and its resources at a specific instant.

Arguments:

state::State: state at instant

total::Float64: total load at instant

selected::Int: selected node at instant; value is zero if load is removed

duration::Float64: duration of all the actions taken during corresponding to the state of this snap

solving_time::Float64: time taken specifically by the solving algorithm (<: AbstractAlgorithm)

instant::Float64

KuMo.State — Type

State

A structure to store the state of the different resources, e.g. nodes and links, during a simulation.

Arguments:

links::SparseMatrixCSC{Float64, Int64}: sparse matrice with the links loads

nodes::SparseVector{Float64, Int64}: sparse vector with the nodes loads

State(n): inner constructor given the number of nodes n

State(links, nodes): inner constructor given the links and nodes of an existing state

KuMo.Topology — Type

Topology{N<:AbstractNode,L<:AbstractLink}

A structure to store the topology of a network. Beside the graph structure itself, it also stores the kinds of all nodes and links.

KuMo.User — Type

User{R<:AbstractRequests}

A structure to store a user information. A user is defined as a location (node id).

KuMo.add_load! — Method

add_load!(state, links, containers, v, n)

Adds load to a given state.

Arguments:

state

links: the load increase to be added on links

containers: the containers load to be added to v

v: node selected to execute a task

n: amount of available nodes

KuMo.capacity — Method

capacity(r::R) where {R<:AbstractResource}

Return the capacity of a resource r.

KuMo.clean — Method

clean(snaps)

Clean the snapshots by merging snaps occurring at the same time.

KuMo.cond_minmax — Method

cond_minmax(x, y, b)

Return the min of x and y if b is false, and the max of x and y otherwise.

KuMo.data — Method

data(location)

Construct data at location. If the location is a collection, a random location is chosen.

KuMo.execution_results — Method

post_simulate(s, snapshots, verbose, output)

Post-simulation process that covers cleaning the snapshots and producing an output.

Arguments:

s: simulated scenario

snapshots: resulting snapshots (before cleaning)

verbose: if set to true, prints information about the output and the snapshots

output: output path

KuMo.execution_results — Method

post_simulate(s, snapshots, verbose, output)

Post-simulation process that covers cleaning the snapshots and producing an output.

Arguments:

s: simulated scenario

snapshots: resulting snapshots (before cleaning)

verbose: if set to true, prints information about the output and the snapshots

output: output path

KuMo.graph — Method

graph(topo::Topology, algorithm::AbstractAlgorithm)

Creates an appropriate digraph using Graph.jl based on a topology and the requirement of an algorithm.

KuMo.init_execution — Method

init_execution(::BatchSimulation)

Initialize a synchronous batch simulation.

KuMo.init_execution — Method

init_execution(::InteractiveRun)

Initialize an interactive run.

KuMo.inner_queue — Function

inner_queue(g, u, j, nodes, capacities, state, ::ShortestPath, ii = 0; lck = ReentrantLock(), demands = nothing, links)

DOCSTRING

Arguments:

g: a graph representing the topology of the network

u: user location

j: requested job

nodes: nodes capacities

capacities: links capacities

state: current state of the network

algo: ShortestPath <: AbstractAlgorithm

ii: a counter to measure the progress in the simulation

lck: a lck for asynchronous simulation

demands: not needed for ShortestPath algorithm

links: description of the links topology

KuMo.insert_sorted! — Function

insert_sorted!(w, val, it = iterate(w))

Insert element in a sorted collection.

Arguments:

w: sorted collection

val: value to be inserted

it: optional iterator

KuMo.job — Method

job(backend::Int, containers::Int, data_location::Int, duration::Float64, frontend::Int)

Method to create new jobs.

KuMo.job_distributions — Method

job_distributions(; backend, container, data_locations, duration, frontend)

Construct a dictionary with random distributions to generate new jobs. Beside data_locations, the other arguments should be a 2-tuple defining normal distributions as in the Distributions.jl package.

Arguments:

backend

container

data_locations: a collection/range of possible data location

duration

frontend

KuMo.make_df — Method

make_df(snapshots::Vector{SnapShot}, topo; verbose = true)

Make a DataFrame from the raw snapshots.

Arguments:

snapshots: A collection of snapshots

topo: topology of the network

verbose: if set to true, it will print a description of the snapshots in the terminal

KuMo.marks — Method

marks(df::DataFrame)

Returns a 4-tuple (a, b, c, d) that marks the start and end of the nodes and links columns in the dataframe.

(a, b) mark the start and end indices of the nodes columns

(c, d) mark the start and end indices of the links columns

KuMo.param — Method

param(r::R) where {R<:AbstractResource}

Default accessor for an optional parameter for R. If no param field exists, returns nothing.

KuMo.pseudo_cost — Function

pseudo_cost(cap, charge, resource, param...)
pseudo_cost(r::<:AbstractResource, charge)

Methods to compute the pseudo-cost of various resources.

KuMo.pseudo_cost — Method

pseudo_cost(r::R, charge) where {R<:AbstractResource}

Compute the pseudo-cost of r given its charge.

KuMo.push_snap! — Method

push_snap!(snapshots, state, total, selected, duration, solving_time, instant, n)

Add a snapshot to an existing collection of snapshots.

Arguments:

snapshots: collection of snapshots

state: current state

total: load

selected: node where a request is executed

duration: duration of the whole resource allocation for the request

solving_time: time taken specifically by the solving algorithm (<: AbstractAlgorithm)

instant: instant when the request is received

n: number of available nodes

KuMo.rand_job — Method

rand_job(jd::Dict)

Create a random job given a job_distribution dictionary.

KuMo.rem_load! — Method

rem_load!(state, links, containers, v, n)

Removes load from a given state.

Arguments:

state

links: the load increase to be removed from links

containers: the containers load to be removed from v

v: node where a task is endind

n: amount of available nodes

KuMo.retrieve_path — Method

retrieve_path(u, v, paths)

Retrieves the path from u to v.

Arguments:

u: source vertex

v: target vertex

paths: list of shortest paths within a network

KuMo.user — Method

user(location)

Construct a user at location. If the location is a collection, a random location is chosen.

KuMo.valid_load — Function

valid_load(s, task, g, capacities, state, algo, demands, ii = 0)

Compute the best load allocation and return if it is a valid one.

Arguments:

s: scenario being simulated

task: task being requested

g: graph of the network topology

capacities: capacities of the network

state: current state of resources

algo: algo used for computing the best allocation cost

demands: if algo is KuMoFlowExt.MinCostFlow, demands are required

ii: a counter to measure the approximative progress of the simulation

KuMo.vtx — Method

vtx(algorithm::AbstractAlgorithm)

Return the number of additional vertices required by the algorithm used to allocate resources in the network.